+++ local/models.py
@@ -1,0 +1,3 @@
+from django.db import models
+from django.utils import timezone
+from productos.models import Preparacion