+++ changelog.md
@@ -5,3 +5,7 @@ Aquí irá el registro de cambios y commits con su fecha y descripción.
 - Creacion de app de productos, la cual involucra la seccion entera de preparacion, ingredientes_preparacion, ingrediente, categoria
 - Añadida la carpeta de documentos con los diagramas de base de datos y de zonas
 - Creada la app core la cual contiene la pagina de home
+
+# 07/04/23 App API Compras
+- Creacion de app de compras, la cual involucra las tablas: Cupon, Compra, Item_compra
+- Creada la App API Local